Create prototype for workload generator using orleans.
This is kind of a first exploration of how to use Orleans for
the workload generation. Right now it is not working
since I am not sure where to start the silo (server process
of orleans) An whether we have to create a orleans client
in the WorkloadGenerator Coordinator to create the grains.

Also the the overall project structure is not clean,
and it will probably need to be designed from scratch again.

I think we generally have the problem of how to seperate
the code we write for the workload generation component
vs. giving the developers using the simulation tool
the possibility to write the individual requests as code
and also implement the data generation logic.