Complete a first working version of workload generation.
Currently just writing the operation results to a log file, in
the future we want to change this to Kafka.

Also no real scheduler in place yet, right now just starting all
transactions asynchronously and then waiting on them.

We probably want a pipeline (mechanism for defining goal throughput)
and ensuring it is reached by scheduler, might also want
to adjust according to how the application that is being
simulated keeps up with the requests.