Start working on workload scheduler using a concurrent queue.
Again this is just a first try at it, but I also got stuck on here
because the executableTransaction (object) inside the queue
cannot be serialized because it has a dictionary containing interfaces
as an attribute. I am bit confused why it is not a problem
when we explicitly pass this dictionary to the grain (as in
the previous version).