Fix bug in OperationService and OperationExecutionService.
We did not pass the response payload to the resolvedOperation, right
now I simply passed the response payload from the input.

Also the ExecutionService did not correctly add the extracted values
to the initial providedValues but instead just returned the extracted
ones.

There is still a bug concerning the DI of the OperationExecution
services.